Moving abort and assert into their own public headers (#6445)
Moving abort and assert into their own public headers
